Inital Commit of Files from previous version
Previous version done for 2019 Fall Semester Class avaliable here:
https://github.com/alexandraburress/488GroupProject

As this is for a different class, I want to make sure that this group has access to this version on my Github, not on the Github of a member no longer associated with this version.
The members from the other group are free to alter that version as they see fit.